Modify seoul256 bg to use default, with exceptions
Note: This is specific to my personal configuration. 

This allows for a tmux configuration that already matches the seoul256 color scheme to switch a vim pane's background color when switching active panes (so it's easier to detect which pane is active).

The exceptions are 'LineNr' and 'Visual*'. LineNr background matches my inactive terminal color, so it looks good when switching between panes, and Visual modes require a separate guibg color anyways.